+++ com.sparta.logistics.delivery/src/main/java/com/sparta/logistics/delivery/repository/DeliveryRouteRepository.java
@@ -1,5 +1,6 @@
-package com.sparta.logistics.delivery.domain;
+package com.sparta.logistics.delivery.repository;
 
+import com.sparta.logistics.delivery.entity.DeliveryRoute;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
